+++ internal/dataplane/sendconfig/sendconfig.go
@@ -4,29 +4,20 @@ import (
 	"bytes"
 	"context"
 	"encoding/hex"
-	"encoding/json"
 	"fmt"
-	"io"
 	"sync"
 	"time"
 
-	"github.com/blang/semver/v4"
-	"github.com/kong/deck/diff"
-	"github.com/kong/deck/dump"
 	"github.com/kong/deck/file"
-	"github.com/kong/deck/state"
-	deckutils "github.com/kong/deck/utils"
 	"github.com/kong/go-kong/kong"
 	"github.com/sirupsen/logrus"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/apimachinery/pkg/types"
 
 	"github.com/kong/kubernetes-ingress-controller/v2/internal/adminapi"
-	"github.com/kong/kubernetes-ingress-controller/v2/internal/dataplane/deckerrors"
 	"github.com/kong/kubernetes-ingress-controller/v2/internal/dataplane/deckgen"
 	"github.com/kong/kubernetes-ingress-controller/v2/internal/dataplane/failures"
 	"github.com/kong/kubernetes-ingress-controller/v2/internal/metrics"
-	"github.com/kong/kubernetes-ingress-controller/v2/internal/versions"
 )
 
 // -----------------------------------------------------------------------------
@@ -34,7 +25,14 @@ import (
 // -----------------------------------------------------------------------------
 
 // PerformUpdate writes `targetContent` to Kong Admin API specified by `kongConfig`.
-func PerformUpdate(ctx context.Context, log logrus.FieldLogger, client *adminapi.Client, config Config, targetContent *file.Content, promMetrics *metrics.CtrlFuncMetrics) ([]byte, []failures.ResourceFailure, error) {
+func PerformUpdate(
+	ctx context.Context,
+	log logrus.FieldLogger,
+	client *adminapi.Client,
+	config Config,
+	targetContent *file.Content,
+	promMetrics *metrics.CtrlFuncMetrics,
+) ([]byte, []failures.ResourceFailure, error) {
 	oldSHA := client.LastConfigSHA()
 	newSHA, err := deckgen.GenerateSHA(targetContent)
 	if err != nil {
@@ -53,175 +51,69 @@ func PerformUpdate(ctx context.Context, log logrus.FieldLogger, client *adminapi
 		}
 	}
 
-	var (
-		metricsProtocol metrics.Protocol
-		parseErr        error
-		resourceErrors  []ResourceError
-	)
+	updateStrategy := ResolveUpdateStrategy(client, config, log)
+
 	timeStart := time.Now()
-	if config.InMemory {
-		metricsProtocol = metrics.ProtocolDBLess
-		err, resourceErrors, parseErr = onUpdateInMemoryMode(ctx, log, targetContent, client.AdminAPIClient())
-	} else {
-		metricsProtocol = metrics.ProtocolDeck
-		dumpConfig := dump.Config{SelectorTags: config.FilterTags, SkipCACerts: config.SkipCACertificates}
-		err = onUpdateDBMode(ctx, targetContent, client, dumpConfig, config)
-	}
+	err, resourceErrors, resourceErrorsParseErr := updateStrategy.Update(ctx, targetContent)
 	duration := time.Since(timeStart)
 
+	metricsProtocol := updateStrategy.MetricsProtocol()
 	if err != nil {
-		dblessFailures := []failures.ResourceFailure{}
-		if parseErr != nil {
-			log.WithError(parseErr).Error("could not parse error response from Kong")
-		} else {
-			for _, ee := range resourceErrors {
-				obj := metav1.PartialObjectMetadata{
-					TypeMeta: metav1.TypeMeta{
-						Kind:       ee.Kind,
-						APIVersion: ee.APIVersion,
-					},
-					ObjectMeta: metav1.ObjectMeta{
-						Namespace: ee.Namespace,
-						Name:      ee.Name,
-						UID:       types.UID(ee.UID),
-					},
-				}
-				for field, problem := range ee.Problems {
-					log.Debug(fmt.Sprintf("adding failure for %s: %s = %s", ee.Name, field, problem))
-					resourceFailure, failureCreateErr := failures.NewResourceFailure(fmt.Sprintf("invalid %s: %s", field, problem), &obj)
-					if failureCreateErr != nil {
-						log.WithError(failureCreateErr).Error("could create resource failure event")
-					} else {
-						dblessFailures = append(dblessFailures, resourceFailure)
-					}
-				}
-			}
-		}
-
+		resourceFailures := resourceErrorsToResourceFailures(resourceErrors, resourceErrorsParseErr, log)
 		promMetrics.RecordPushFailure(metricsProtocol, duration, client.BaseRootURL(), err)
-		return nil, dblessFailures, err
+		return nil, resourceFailures, err
 	}
 
 	promMetrics.RecordPushSuccess(metricsProtocol, duration, client.BaseRootURL())
-	log.Info("successfully synced configuration to kong.")
-	return newSHA, []failures.ResourceFailure{}, nil
+	log.Info("successfully synced configuration to kong")
+	return newSHA, nil, nil
 }
 
 // -----------------------------------------------------------------------------
 // Sendconfig - Private Functions
 // -----------------------------------------------------------------------------
 
-type InMemoryClient interface {
-	BaseRootURL() string
-	ReloadDeclarativeRawConfig(ctx context.Context, config io.Reader, checkHash bool, flattenErrors bool) ([]byte, error)
+type KonnectAwareClient interface {
+	IsKonnect() bool
 }
 
-// onUpdateInMemoryMode takes a context, logger, deck state file struct, and a Kong client. It POSTs the state to
-// the Kong DB-less /config endpoint, setting that Kong instance's configuration to the configuration in the state.
-// It returns an error, a slice of ResourceErrors, and a second error. The first error indicates any overall issues
-// POSTing configuration to the Kong instance. If the first error is not nil, the ResourceError slice contains a
-// ResourceError for each Kubernetes resource that generated invalid Kong configuration. The second error is non-nil if
-// onUpdateInMemoryMode could not parse Kong's error response into a slice of ResourceError, and indicates why parsing
-// failed if so.
-func onUpdateInMemoryMode(
+type StatusClient interface {
+	Status(context.Context) (*kong.Status, error)
+}
+
+// hasConfigurationChanged verifies whether configuration has changed by comparing old and new config's SHAs.
+// In case the SHAs are equal, it still can return true if a client is considered crashed based on its status.
+func hasConfigurationChanged(
 	ctx context.Context,
+	oldSHA, newSHA []byte,
+	client KonnectAwareClient,
+	statusClient StatusClient,
 	log logrus.FieldLogger,
-	state *file.Content,
-	client InMemoryClient,
-) (error, []ResourceError, error) {
-	// Kong will error out if this is set
-	state.Info = nil
-	// Kong errors out if `null`s are present in `config` of plugins
-	deckgen.CleanUpNullsInPluginConfigs(state)
-	var parseErr error
-	var resourceErrors []ResourceError
-	var errBody []byte
+) (bool, error) {
+	if !bytes.Equal(oldSHA, newSHA) {
+		return true, nil
+	}
+	if !hasSHAUpdateAlreadyBeenReported(newSHA) {
+		log.Debugf("sha %s has been reported", hex.EncodeToString(newSHA))
+	}
+	// In case of Konnect, we skip further steps that are meant to detect Kong instances crash/reset
+	// that are not relevant for Konnect.
+	// We're sure that if oldSHA and newSHA are equal, we are safe to skip the update.
+	if client.IsKonnect() {
+		return false, nil
+	}
 
-	config, err := json.Marshal(state)
+	// Check if a Kong instance has no configuration yet (could mean it crashed, was rebooted, etc.).
+	hasNoConfiguration, err := kongHasNoConfiguration(ctx, statusClient, log)
 	if err != nil {
-		return fmt.Errorf("constructing kong configuration: %w", err), resourceErrors, parseErr
+		return false, fmt.Errorf("failed to verify kong readiness: %w", err)
+	}
+	// Kong instance has no configuration, we should push despite the oldSHA and newSHA being equal.
+	if hasNoConfiguration {
+		return true, nil
 	}
 
-	var flattened bool
-	if !versions.GetKongVersion().MajorMinorOnly().LTE(versions.FlattenedErrorCutoff) {
-		// Kong's API library combines KVs in the request body (the config) and query string (check hash, flattened)
-		// into a single set of parameters: https://github.com/Kong/go-kong/pull/271#issuecomment-1416212852
-		// KIC therefore must _not_ request flattened errors on versions that do not support it, as otherwise Kong
-		// will interpret the query string toggle as part of the config, and will reject it, as "flattened_errors" is
-		// not a valid config key. KIC only sends this query parameter if Kong is 3.2 or higher.
-		flattened = true
-	}
-
-	log.Debug("sending configuration to Kong Admin API")
-	if errBody, err = client.ReloadDeclarativeRawConfig(ctx, bytes.NewReader(config), true, flattened); err != nil {
-		resourceErrors, parseErr = parseFlatEntityErrors(errBody, log)
-		return err, resourceErrors, parseErr
-	}
-
-	return nil, resourceErrors, parseErr
-}
-
-func onUpdateDBMode(
-	ctx context.Context,
-	targetContent *file.Content,
-	client *adminapi.Client,
-	dumpConfig dump.Config,
-	config Config,
-) error {
-	cs, err := currentState(ctx, client, dumpConfig)
-	if err != nil {
-		return fmt.Errorf("failed getting current state for %s: %w", client.BaseRootURL(), err)
-	}
-
-	ts, err := targetState(ctx, targetContent, cs, config.Version, client, dumpConfig)
-	if err != nil {
-		return deckerrors.ConfigConflictError{Err: err}
-	}
-
-	syncer, err := diff.NewSyncer(diff.SyncerOpts{
-		CurrentState:    cs,
-		TargetState:     ts,
-		KongClient:      client.AdminAPIClient(),
-		SilenceWarnings: true,
-	})
-	if err != nil {
-		return fmt.Errorf("creating a new syncer for %s: %w", client.BaseRootURL(), err)
-	}
-
-	_, errs := syncer.Solve(ctx, config.Concurrency, false)
-	if errs != nil {
-		return deckutils.ErrArray{Errors: errs}
-	}
-
-	return nil
-}
-
-func currentState(ctx context.Context, kongClient *adminapi.Client, dumpConfig dump.Config) (*state.KongState, error) {
-	rawState, err := dump.Get(ctx, kongClient.AdminAPIClient(), dumpConfig)
-	if err != nil {
-		return nil, fmt.Errorf("loading configuration from kong: %w", err)
-	}
-
-	return state.Get(rawState)
-}
-
-func targetState(
-	ctx context.Context,
-	targetContent *file.Content,
-	currentState *state.KongState,
-	version semver.Version,
-	kongClient *adminapi.Client,
-	dumpConfig dump.Config,
-) (*state.KongState, error) {
-	rawState, err := file.Get(ctx, targetContent, file.RenderConfig{
-		CurrentState: currentState,
-		KongVersion:  version,
-	}, dumpConfig, kongClient.AdminAPIClient())
-	if err != nil {
-		return nil, err
-	}
-
-	return state.Get(rawState)
+	return false, nil
 }
 
 var (
@@ -250,46 +142,6 @@ func hasSHAUpdateAlreadyBeenReported(latestUpdateSHA []byte) bool {
 	return false
 }
 
-type KonnectAwareClient interface {
-	IsKonnect() bool
-}
-
-type StatusClient interface {
-	Status(context.Context) (*kong.Status, error)
-}
-
-// hasConfigurationChanged verifies whether configuration has changed by comparing old and new config's SHAs.
-// In case the SHAs are equal, it still can return true if a client is considered crashed based on its status.
-func hasConfigurationChanged(
-	ctx context.Context,
-	oldSHA, newSHA []byte,
-	client KonnectAwareClient,
-	statusClient StatusClient,
-	log logrus.FieldLogger,
-) (bool, error) {
-	if !bytes.Equal(oldSHA, newSHA) {
-		return true, nil
-	}
-	if !hasSHAUpdateAlreadyBeenReported(newSHA) {
-		log.Debugf("sha %s has been reported", hex.EncodeToString(newSHA))
-	}
-	// In case of Konnect, we skip further steps as it doesn't report its configuration hash.
-	if client.IsKonnect() {
-		return false, nil
-	}
-
-	hasNoConfiguration, err := kongHasNoConfiguration(ctx, statusClient, log)
-	if err != nil {
-		return false, fmt.Errorf("failed to verify kong readiness: %w", err)
-	}
-	// Kong instance has no configuration, we should push despite the oldSHA and newSHA being equal.
-	if hasNoConfiguration {
-		return true, nil
-	}
-
-	return false, nil
-}
-
 const wellKnownInitialHash = "00000000000000000000000000000000"
 
 // kongHasNoConfiguration checks Kong's status endpoint and read its config hash.
@@ -308,4 +160,42 @@ func kongHasNoConfiguration(ctx context.Context, client StatusClient, log logrus
 	}
 
 	return false, nil
+}
+
+// resourceErrorsToResourceFailures translates a slice of ResourceError to a slice of failures.ResourceFailure.
+// In case of parseErr being not nil, it just returns a nil slice.
+func resourceErrorsToResourceFailures(resourceErrors []ResourceError, parseErr error, log logrus.FieldLogger) []failures.ResourceFailure {
+	if parseErr != nil {
+		log.WithError(parseErr).Error("failed parsing resource errors")
+		return nil
+	}
+
+	var out []failures.ResourceFailure
+	for _, ee := range resourceErrors {
+		obj := metav1.PartialObjectMetadata{
+			TypeMeta: metav1.TypeMeta{
+				Kind:       ee.Kind,
+				APIVersion: ee.APIVersion,
+			},
+			ObjectMeta: metav1.ObjectMeta{
+				Namespace: ee.Namespace,
+				Name:      ee.Name,
+				UID:       types.UID(ee.UID),
+			},
+		}
+		for field, problem := range ee.Problems {
+			log.Debug(fmt.Sprintf("adding failure for %s: %s = %s", ee.Name, field, problem))
+			resourceFailure, failureCreateErr := failures.NewResourceFailure(
+				fmt.Sprintf("invalid %s: %s", field, problem),
+				&obj,
+			)
+			if failureCreateErr != nil {
+				log.WithError(failureCreateErr).Error("could create resource failure event")
+			} else {
+				out = append(out, resourceFailure)
+			}
+		}
+	}
+
+	return out
 }