+++ test/internal/testenv/feature_gates.go
@@ -1,0 +1,10 @@
+//go:build !knative
+// +build !knative
+
+package testenv
+
+import "github.com/kong/kubernetes-ingress-controller/v2/test/consts"
+
+func getFeatureGates() string {
+	return consts.DefaultFeatureGates
+}